refactor: organize code for better code structure
