+++ src/dev_requirements.txt
@@ -1,0 +1,3 @@
+-r requirements.txt
+pytest==8.2.2
+pytest-django==4.8.0
