+++ src/django_app/urls.py
@@ -17,7 +17,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
-from api.receivers.infra.django_ninja_app.urls import api
+from receivers.infra.django_ninja_app.urls import api
 
 urlpatterns = [
     path('admin/', admin.site.urls),
