+++ src/receivers/infra/django_ninja_app/models.py
